restructure dependency module, fix test, update podspec and package.swift
