--- Sources/Dependencies/Internal/Dependency+TypeName.swift
@@ -1,15 +1,0 @@
-@usableFromInline
-func typeName(_ type: Any.Type) -> String {
-    var name = _typeName(type, qualified: true)
-    if let index = name.firstIndex(of: ".") {
-        name.removeSubrange(...index)
-    }
-    let sanitizedName =
-        name
-        .replacingOccurrences(
-            of: #"\(unknown context at \$[[:xdigit:]]+\)\."#,
-            with: "",
-            options: .regularExpression
-        )
-    return sanitizedName
-}
