+++ Sources/RxComposableArchitecture/Dependencies/Internal/Dependencies+TypeName.swift
@@ -1,0 +1,15 @@
+@usableFromInline
+func dependencyTypeName(_ type: Any.Type) -> String {
+  var name = _typeName(type, qualified: true)
+  if let index = name.firstIndex(of: ".") {
+    name.removeSubrange(...index)
+  }
+  let sanitizedName =
+    name
+    .replacingOccurrences(
+      of: #"\(unknown context at \$[[:xdigit:]]+\)\."#,
+      with: "",
+      options: .regularExpression
+    )
+  return sanitizedName
+}
