+++ Sources/RxComposableArchitecture/Export.swift
@@ -1,2 +1,1 @@
 @_exported import CasePaths
-@_exported import XCTestDynamicOverlay
