+++ Tests/RxComposableArchitectureTests/DependencyTests/DependencyValuesTests.swift
@@ -1,4 +1,3 @@
-import Dependencies
 import RxComposableArchitecture
 import XCTest
 
@@ -27,27 +26,27 @@ final class DependencyValuesTests: XCTestCase {
                 }
             } issueMatcher: {
                 $0.compactDescription == """
-                    "@Dependency(\\.missingLiveDependency)" has no live implementation, but was accessed \
-                    from a live context.
-
-                      Location:
-                        DependenciesTests/DependencyValuesTests.swift:\(line)
-                      Key:
-                        TestKey
-                      Value:
-                        Int
-
-                    Every dependency registered with the library must conform to "DependencyKey", and that \
-                    conformance must be visible to the running application.
-
-                    To fix, make sure that "TestKey" conforms to "DependencyKey" by providing a live \
-                    implementation of your dependency, and make sure that the conformance is linked with \
-                    this current application.
-                    """
+              "@Dependency(\\.missingLiveDependency)" has no live implementation, but was accessed \
+              from a live context.
+              
+                Location:
+                  RxComposableArchitectureTests/DependencyValuesTests.swift:\(line)
+                Key:
+                  TestKey
+                Value:
+                  Int
+              
+              Every dependency registered with the library must conform to "DependencyKey", and that \
+              conformance must be visible to the running application.
+              
+              To fix, make sure that "TestKey" conforms to "DependencyKey" by providing a live \
+              implementation of your dependency, and make sure that the conformance is linked with \
+              this current application.
+              """
             }
         #endif
     }
-
+    
     func testWithValues() {
         let date = DependencyValues.withValues {
             $0.date = .constant(someDate)
@@ -55,25 +54,25 @@ final class DependencyValuesTests: XCTestCase {
             @Dependency(\.date) var date
             return date.now
         }
-
+        
         let defaultDate = DependencyValues.withValues {
             $0.context = .live
         } operation: { () -> Date in
             @Dependency(\.date) var date
             return date.now
         }
-
+        
         XCTAssertEqual(date, someDate)
         XCTAssertNotEqual(defaultDate, someDate)
     }
-
+    
     func testWithValue() {
         DependencyValues.withValue(\.context, .live) {
             let date = DependencyValues.withValue(\.date, .constant(someDate)) { () -> Date in
                 @Dependency(\.date) var date
                 return date.now
             }
-
+            
             XCTAssertEqual(date, someDate)
             XCTAssertNotEqual(DependencyValues._current.date.now, someDate)
         }
